@@ -1,8 +1,8 @@
-// swift-tools-version:4.0
+// swift-tools-version:5.0
 // Package.swift
 /*
  The MIT License (MIT)
- Copyright (c) 2016 Eric Marchand (phimage)
+ Copyright (c) 2019 Eric Marchand (phimage) & Dongri Jin <dongriat@gmail.com>
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
@@ -27,8 +27,18 @@ let package = Package(
     products: [
         .library(name: "OAuthSwift", targets: ["OAuthSwift"]),
     ],
+    dependencies: [
+        .package(url: "https://github.com/phimage/Erik.git", .upToNextMajor(from: "5.0.0")),
+        .package(url: "https://github.com/tid-kijyun/Kanna.git", .upToNextMajor(from: "5.2.4")),
+        .package(url: "https://github.com/httpswift/swifter.git", .upToNextMajor(from: "1.5.0"))
+    ],
     targets: [
         .target(name: "OAuthSwift", dependencies: [], path: "Sources"),
-        .testTarget(name: "OAuthSwiftTests", dependencies: ["OAuthSwift"], path: "OAuthSwiftTests"),
+        .testTarget(name: "OAuthSwiftTests", dependencies: ["OAuthSwift", "Erik", "Kanna", "Swifter"], path: "OAuthSwiftTests"),
     ]
 )
+
+#if os(Linux)
+package.dependencies.append(.package(url: "https://github.com/apple/swift-crypto.git", from: "1.0.0"))
+package.targets[0].dependencies.append("Crypto")
+#endif
