@@ -9,10 +9,10 @@
 import OAuthSwift
 
 #if os(iOS)
-    import UIKit
-    import SafariServices
+import UIKit
+import SafariServices
 #elseif os(OSX)
-    import AppKit
+import AppKit
 #endif
 
 class ViewController: OAuthViewController {
@@ -21,13 +21,13 @@ class ViewController: OAuthViewController {
     
     var currentParameters = [String: String]()
     let formData = Semaphore<FormViewControllerData>()
-
+    
     lazy var internalWebViewController: WebViewController = {
         let controller = WebViewController()
         #if os(OSX)
-            controller.view = NSView(frame: NSRect(x:0, y:0, width: 450, height: 500)) // needed if no nib or not loaded from storyboard
+        controller.view = NSView(frame: NSRect(x:0, y:0, width: 450, height: 500)) // needed if no nib or not loaded from storyboard
         #elseif os(iOS)
-            controller.view = UIView(frame: UIScreen.main.bounds) // needed if no nib or not loaded from storyboard
+        controller.view = UIView(frame: UIScreen.main.bounds) // needed if no nib or not loaded from storyboard
         #endif
         controller.delegate = self
         controller.viewDidLoad() // allow WebViewController to use this ViewController as parent to be presented
@@ -85,7 +85,7 @@ extension ViewController {
             // Cancel
             return
         }
-
+        
         parameters["consumerKey"] = data.key
         parameters["consumerSecret"] = data.secret
         
@@ -96,9 +96,9 @@ extension ViewController {
                 self.showAlertView(title: "Key and secret must not be empty", message: message)
             }
         }
-
+        
         parameters["name"] = service
-
+        
         switch service {
         case "Imgur" :
             doOAuthImgur(parameters)
@@ -172,6 +172,10 @@ extension ViewController {
             doOAuthNoun(parameters)
         case "Lyft":
             doOAuthLyft(parameters)
+        case "Twitch":
+            doOAuthTwitch(parameters)
+        case "Reddit":
+            doOauthReddit(parameters)
         default:
             print("\(service) not implemented")
         }
@@ -190,14 +194,14 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/500px")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/500px")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Spotify
@@ -216,16 +220,16 @@ extension ViewController {
         let _ = oauthswift.authorize(
             withCallbackURL: URL(string: "oauth-swift://oauth-callback/spotify")!,
             scope: "user-library-modify",
-            state: state,
-            success: { credential, response, parameters in
-                self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
-                print(error.description)
-            }
-        )
+            state: state) { result in
+                switch result {
+                case .success(let (credential, _, _)):
+                    self.showTokenAlert(name: serviceParameters["name"], credential: credential)
+                case .failure(let error):
+                    print(error.description)
+                }
+        }
     }
-
+    
     // MARK: Imgur
     func doOAuthImgur(_ serviceParameters: [String:String]){
         let oauthswift = OAuth2Swift(
@@ -240,20 +244,20 @@ extension ViewController {
         oauthswift.encodeCallbackURLQuery = false
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
-
+        
         let _ = oauthswift.authorize(
             withCallbackURL: URL(string: "oauth-swift://oauth-callback/imgur")!,
             scope: "",
-            state: state,
-            success: { credential, response, parameters in
-                self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
-                print(error.description)
-            }
-        )
+            state: state) { result in
+                switch result {
+                case .success(let (credential, _, _)):
+                    self.showTokenAlert(name: serviceParameters["name"], credential: credential)
+                case .failure(let error):
+                    print(error.description)
+                }
+        }
     }
-
+    
     // MARK: Twitter
     func doOAuthTwitter(_ serviceParameters: [String:String]){
         let oauthswift = OAuth1Swift(
@@ -266,28 +270,28 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "http://oauthswift.herokuapp.com/callback/twitter")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "http://oauthswift.herokuapp.com/callback/twitter")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testTwitter(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testTwitter(_ oauthswift: OAuth1Swift) {
-        let _ = oauthswift.client.get(
-            "https://api.twitter.com/1.1/statuses/mentions_timeline.json", parameters: [:],
-            success: { response in
+        let _ = oauthswift.client.get("https://api.twitter.com/1.1/statuses/mentions_timeline.json", parameters: [:]) { result in
+            switch result {
+            case .success(let response):
                 let jsonDict = try? response.jsonObject()
                 print(String(describing: jsonDict))
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
-
+    
     // MARK: Flickr
     func doOAuthFlickr(_ serviceParameters: [String:String]){
         let oauthswift = OAuth1Swift(
@@ -300,16 +304,17 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/flickr")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/flickr")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testFlickr(oauthswift, consumerKey: serviceParameters["consumerKey"]!)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
+    
     func testFlickr (_ oauthswift: OAuth1Swift, consumerKey: String) {
         let url :String = "https://api.flickr.com/services/rest/"
         let parameters :Dictionary = [
@@ -320,16 +325,15 @@ extension ViewController {
             "nojsoncallback" : "1",
             "extras"         : "url_q,url_z"
         ]
-        let _ = oauthswift.client.get(
-            url, parameters: parameters,
-            success: { response in
+        let _ = oauthswift.client.get(url, parameters: parameters) { result in
+            switch result {
+            case .success(let response):
                 let jsonDict = try? response.jsonObject()
                 print(jsonDict as Any)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Github
@@ -345,14 +349,14 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/github")!, scope: "user,repo", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/github")!, scope: "user,repo", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     //MARK: Salesforce
@@ -368,14 +372,14 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/salesforce")!, scope: "full", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/salesforce")!, scope: "full", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Instagram
@@ -394,30 +398,29 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/instagram")!, scope: "likes+comments", state:state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/instagram")!, scope: "likes+comments", state:state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testInstagram(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
+    
     func testInstagram(_ oauthswift: OAuth2Swift) {
         let url :String = "https://api.instagram.com/v1/users/1574083/?access_token=\(oauthswift.client.credential.oauthToken)"
         let parameters :Dictionary = Dictionary<String, AnyObject>()
-        let _ = oauthswift.client.get(
-            url, parameters: parameters,
-            success: { response in
+        let _ = oauthswift.client.get(url, parameters: parameters) { result in
+            switch result {
+            case .success(let response):
                 let jsonDict = try? response.jsonObject()
                 print(jsonDict as Any)
-                
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Foursquare
@@ -432,14 +435,14 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/foursquare")!, scope: "", state: "",
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/foursquare")!, scope: "", state: "") { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Fitbit
@@ -455,14 +458,14 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/fitbit")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/fitbit")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     func doOAuthFitbit2(_ serviceParameters: [String:String]) {
@@ -479,28 +482,28 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/fitbit2")!, scope: "profile weight", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/fitbit2")!, scope: "profile weight", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testFitbit2(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testFitbit2(_ oauthswift: OAuth2Swift) {
         let _ = oauthswift.client.get(
             "https://api.fitbit.com/1/user/-/profile.json",
-            parameters: [:],
-            success: { response in
-                let jsonDict = try? response.jsonObject()
-                print(jsonDict as Any)
-            },
-            failure: { error in
-                print(error.description)
-            }
-        )
+            parameters: [:]) { result in
+                switch result {
+                case .success(let response):
+                    let jsonDict = try? response.jsonObject()
+                    print(jsonDict as Any)
+                case .failure(let error):
+                    print(error.description)
+                }
+        }
     }
     
     // MARK: Withings
@@ -516,27 +519,29 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/withings")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/withings")!) { result in
+            switch result {
+            case .success(let (credential, _, parameters)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testWithings(oauthswift, userId: parameters["userid"] as! String)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testWithings(_ oauthswift: OAuth1Swift, userId : String) {
         oauthswift.client.paramsLocation = .requestURIQuery
         let _ = oauthswift.client.get(
-            "https://wbsapi.withings.net/v2/measure", parameters: ["action":"getactivity", "userid":userId, "date":"2016-02-15"],
-            success: { response in
-                let jsonDict = try? response.jsonObject()
-                print(jsonDict as Any)
-            }, failure: { error in
-                print(error.description)
-            }
-        )
+            "https://wbsapi.withings.net/v2/measure",
+            parameters: ["action": "getactivity", "userid": userId, "date":"2016-02-15"]) { result in
+                switch result {
+                case .success(let response):
+                    let jsonDict = try? response.jsonObject()
+                    print(jsonDict as Any)
+                case .failure(let error):
+                    print(error.description)
+                }
+        }
     }
     
     // MARK: Linkedin
@@ -552,26 +557,27 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/linkedin")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/linkedin")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testLinkedin(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testLinkedin(_ oauthswift: OAuth1Swift) {
         let _ = oauthswift.client.get(
-            "https://api.linkedin.com/v1/people/~", parameters: [:],
-            success: { response in
+        "https://api.linkedin.com/v1/people/~", parameters: [:]) { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     func doOAuthLinkedin2(_ serviceParameters: [String:String]){
@@ -586,28 +592,28 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "http://oauthswift.herokuapp.com/callback/linkedin2")!, scope: "r_fullprofile", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "http://oauthswift.herokuapp.com/callback/linkedin2")!, scope: "r_fullprofile", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testLinkedin2(oauthswift)
                 
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testLinkedin2(_ oauthswift: OAuth2Swift) {
         let _ = oauthswift.client.get(
-            "https://api.linkedin.com/v1/people/~?format=json", parameters: [:],
-            success: { response in
+        "https://api.linkedin.com/v1/people/~?format=json", parameters: [:]) { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Smugmug
@@ -625,14 +631,14 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         // NOTE: Smugmug's callback URL is configured on their site and the one passed in is ignored.
         let _ = oauthswift.authorize(
-            withCallbackURL: "oauth-swift://oauth-callback/smugmug",
-            success: { credential, response, parameters in
+        withCallbackURL: "oauth-swift://oauth-callback/smugmug") { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Dropbox
@@ -640,32 +646,33 @@ extension ViewController {
         let oauthswift = OAuth2Swift(
             consumerKey:    serviceParameters["consumerKey"]!,
             consumerSecret: serviceParameters["consumerSecret"]!,
-            authorizeUrl:   "https://www.dropbox.com/1/oauth2/authorize",
-            accessTokenUrl: "https://api.dropbox.com/1/oauth2/token",
+            authorizeUrl:   "https://www.dropbox.com/oauth2/authorize",
+            accessTokenUrl: "https://api.dropbox.com/oauth2/token",
             responseType:   "token"
         )
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/dropbox")!, scope: "", state: "",
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/dropbox")!, scope: "", state: "") { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 // Get Dropbox Account Info
                 let parameters =  Dictionary<String, AnyObject>()
                 let _ = oauthswift.client.get(
-                    "https://api.dropbox.com/1/account/info", parameters: parameters,
-                    success: { response in
+                "https://api.dropbox.com/2/users/get_account", parameters: parameters) { result in
+                    switch result {
+                    case .success(let response):
                         let jsonDict = try? response.jsonObject()
                         print(jsonDict as Any)
-                    }, failure: { error in
+                    case .failure(let error):
                         print(error)
                     }
-                )
-            },
-            failure: { error in
+                }
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Dribbble
@@ -680,26 +687,25 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/dribbble")!, scope: "", state: "",
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/dribbble")!, scope: "", state: "") { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 // Get User
                 let parameters =  [String: Any]()
-                let _ = oauthswift.client.get(
-                    "https://api.dribbble.com/v1/user?access_token=\(credential.oauthToken)", parameters: parameters,
-                    success: { response in
+                let _ = oauthswift.client.get("https://api.dribbble.com/v1/user?access_token=\(credential.oauthToken)", parameters: parameters) { result in
+                    switch result {
+                    case .success(let response):
                         let jsonDict = try? response.jsonObject()
                         print(jsonDict as Any)
-                    },
-                    failure: { error in
+                    case .failure(let error):
                         print(error)
                     }
-                )
-            },
-            failure: { error in
+                }
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: BitBucket
@@ -714,27 +720,27 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/bitbucket")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/bitbucket")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testBitBucket(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testBitBucket(_ oauthswift: OAuth1Swift) {
         let _ = oauthswift.client.get(
-            "https://bitbucket.org/api/1.0/user", parameters: [:],
-            success: { response in
+        "https://bitbucket.org/api/1.0/user", parameters: [:]) { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Google
@@ -752,26 +758,27 @@ extension ViewController {
         oauthswift.allowMissingStateCheck = true
         // in plist define a url schem with: your.bundle.id:
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/google")!, scope: "https://www.googleapis.com/auth/drive", state: "",
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/google")!, scope: "https://www.googleapis.com/auth/drive", state: "") { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 let parameters =  Dictionary<String, AnyObject>()
                 // Multi-part upload
                 let _ = oauthswift.client.postImage(
-                    "https://www.googleapis.com/upload/drive/v2/files", parameters: parameters, image: self.snapshot(),
-                    success: { response in
+                "https://www.googleapis.com/upload/drive/v2/files", parameters: parameters, image: self.snapshot()) { result in
+                    switch result {
+                    case .success(let response):
                         let jsonDict = try? response.jsonObject()
                         print("SUCCESS: \(String(describing: jsonDict))")
-                    },
-                    failure: { error in
+                    case .failure(let error):
                         print(error)
                     }
-                )
-            },
-            failure: { error in
+                    
+                }
+            case .failure(let error):
                 print("ERROR: \(error.localizedDescription)")
             }
-        )
+        }
     }
     
     // MARK/ Intuit
@@ -787,67 +794,65 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/intuit")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/intuit")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testIntuit(oauthswift, serviceParameters: serviceParameters)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testIntuit(_ oauthswift: OAuth1Swift, serviceParameters: [String:String]){
-        if let companyId = serviceParameters["companyId"]  {
-            let _ = oauthswift.client.get(
-                "https://sandbox-quickbooks.api.intuit.com/v3/company/\(companyId)/account/1", headers: ["Accept":"application/json"],
-                success: { response in
-                    if let jsonDict = try? response.jsonObject(options: .allowFragments) , let dico = jsonDict as? [String: AnyObject] {
-                        print(dico)
-                        
-                        // XXX to generate with good date etc...
-                        let jsonUpdate: [String : Any] = [
-                            "Name": "Accounts Payable (A/P)",
-                            "SubAccount": false,
-                            "FullyQualifiedName": "Accounts Payable (A/P)",
-                            "Active": true,
-                            "Classification": "Liability",
-                            "Description": "Description added during update.",
-                            "AccountType": "Accounts Payable",
-                            "AccountSubType": "AccountsPayable",
-                            "CurrentBalance": -1091.23,
-                            "CurrentBalanceWithSubAccounts": -1091.23,
-                            "domain": "QBO",
-                            "sparse": false,
-                            "Id": "33",
-                            "SyncToken": "0",
-                            "MetaData": [
-                                "CreateTime": "2014-09-12T10:12:02-07:00",
-                                "LastUpdatedTime": "2015-06-30T15:09:07-07:00"
-                            ]
+        guard let companyId = serviceParameters["companyId"] else { return }
+        
+        let _ = oauthswift.client.get("https://sandbox-quickbooks.api.intuit.com/v3/company/\(companyId)/account/1", headers: ["Accept":"application/json"]) { result in
+            switch result {
+            case .success(let response):
+                if let jsonDict = try? response.jsonObject(options: .allowFragments) , let dico = jsonDict as? [String: AnyObject] {
+                    print(dico)
+                    
+                    // XXX to generate with good date etc...
+                    let jsonUpdate: [String : Any] = [
+                        "Name": "Accounts Payable (A/P)",
+                        "SubAccount": false,
+                        "FullyQualifiedName": "Accounts Payable (A/P)",
+                        "Active": true,
+                        "Classification": "Liability",
+                        "Description": "Description added during update.",
+                        "AccountType": "Accounts Payable",
+                        "AccountSubType": "AccountsPayable",
+                        "CurrentBalance": -1091.23,
+                        "CurrentBalanceWithSubAccounts": -1091.23,
+                        "domain": "QBO",
+                        "sparse": false,
+                        "Id": "33",
+                        "SyncToken": "0",
+                        "MetaData": [
+                            "CreateTime": "2014-09-12T10:12:02-07:00",
+                            "LastUpdatedTime": "2015-06-30T15:09:07-07:00"
                         ]
-                        
-                        // FIXME #80
-                        let _ = oauthswift.client.post(
-                            "https://sandbox-quickbooks.api.intuit.com/v3/company/\(companyId)/account?operation=update", parameters: jsonUpdate,
-                            headers: ["Accept": "application/json", "Content-Type":"application/json"],
-                            success: { response in
+                    ]
+                    
+                    // FIXME #80
+                    let _ = oauthswift.client.post(
+                        "https://sandbox-quickbooks.api.intuit.com/v3/company/\(companyId)/account?operation=update", parameters: jsonUpdate,
+                        headers: ["Accept": "application/json", "Content-Type":"application/json"]) { result in
+                            switch result {
+                            case .success(let response):
                                 print(response.data)
-                            },
-                            failure: { error in
+                            case .failure(let error):
                                 print(error)
                             }
-                        )
                     }
-                    else {
-                        print("no json response")
-                    }
-                    
-                },
-                failure: { error in
-                    print(error)
                 }
-            )
+                else {
+                    print("no json response")
+                }
+            case .failure(let error):
+                print(error)
+            }
         }
     }
     
@@ -864,14 +869,14 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/zaim")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/zaim")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Tumblr
@@ -887,43 +892,43 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/tumblr")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/tumblr")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testTumblr(oauthswift, serviceParameters: serviceParameters)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     func testTumblr(_ oauthswift: OAuth1Swift, serviceParameters: [String:String]){
         let _ = oauthswift.client.get(
-            "https://api.tumblr.com/v2/user/info", headers: ["Accept":"application/json"],
-            success: { response in
+        "https://api.tumblr.com/v2/user/info", headers: ["Accept":"application/json"]) { result in
+            switch result {
+            case .success(let response):
                 if let jsonDict = try? response.jsonObject(options: .allowFragments) , let dico = jsonDict as? [String: Any] {
                     print(dico)
                 }
                 else {
                     print("no json response")
                 }
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
         
         let url = serviceParameters["blogURL"] ?? "good.tumblr.com"
         let _ = oauthswift.client.post(
-            "https://api.tumblr.com/v2/user/follow", parameters: ["url": url],
-            success: { response in
+        "https://api.tumblr.com/v2/user/follow", parameters: ["url": url]) { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     
@@ -941,14 +946,14 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/slack")!, scope: "", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/slack")!, scope: "", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     
     // MARK: Uber
@@ -967,14 +972,14 @@ extension ViewController {
         let state = generateState(withLength: 20)
         let redirectURL = "https://oauthswift.herokuapp.com/callback/uber".addingPercentEncoding(withAllowedCharacters: CharacterSet.urlHostAllowed)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: redirectURL!)!, scope: "profile", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: redirectURL!)!, scope: "profile", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     
     // MARK Gitter
@@ -991,14 +996,14 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/gitter")!, scope: "flow", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/gitter")!, scope: "flow", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     
     // MAK : Facebook
@@ -1015,25 +1020,27 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/facebook")!, scope: "public_profile", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/facebook")!, scope: "public_profile", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testFacebook(oauthswift)
-            }, failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     func testFacebook(_ oauthswift: OAuth2Swift) {
         let _ = oauthswift.client.get(
-            "https://graph.facebook.com/me?",
-            success: { response in
+        "https://graph.facebook.com/me?") { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Hatena
@@ -1049,14 +1056,14 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/hatena")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/hatena")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     // MARK: Trello
@@ -1072,27 +1079,28 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/trello")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/trello")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testTrello(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     
     func testTrello(_ oauthswift: OAuth1Swift) {
         let _ = oauthswift.client.get(
-            "https://trello.com/1/members/me/boards",
-            success: { response in
+        "https://trello.com/1/members/me/boards") { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Buffer
@@ -1108,27 +1116,28 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/buffer")!, scope: "", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/buffer")!, scope: "", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testBuffer(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     
     func testBuffer(_ oauthswift: OAuth2Swift) {
         let _ = oauthswift.client.get(
-            "https://api.bufferapp.com/1/user.json",
-            success: { response in
+        "https://api.bufferapp.com/1/user.json") { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Goodreads
@@ -1146,29 +1155,29 @@ extension ViewController {
         // The callback url you set here doesn't seem to make a differnce,
         // you have to set it up at the site when you get your developer key.
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/goodreads")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/goodreads")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testGoodreads(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
-        
+        }
     }
     
     func testGoodreads(_ oauthswift: OAuth1Swift) {
         let _ = oauthswift.client.get(
-            "https://www.goodreads.com/api/auth_user",
-            success: { response in
+        "https://www.goodreads.com/api/auth_user") { result in
+            switch result {
+            case .success(let response):
                 // Most Goodreads methods return XML, you'll need a way to parse it.
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Typetalk
@@ -1185,27 +1194,28 @@ extension ViewController {
         self.oauthswift = oauthswift
         oauthswift.authorizeURLHandler = getURLHandler()
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/typetalk")!, scope: "", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/typetalk")!, scope: "", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testTypetalk(oauthswift)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
+        }
     }
     
     func testTypetalk(_ oauthswift: OAuth2Swift) {
         let _ = oauthswift.client.get(
-            "https://typetalk.in/api/v1/profile",
-            success: { response in
+        "https://typetalk.in/api/v1/profile") { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: SoundCloud
@@ -1221,27 +1231,28 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/soundcloud")!, scope: "", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "https://oauthswift.herokuapp.com/callback/soundcloud")!, scope: "", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
                 self.testSoundCloud(oauthswift,credential.oauthToken)
-            },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
             }
-        )
+        }
     }
     
     func testSoundCloud(_ oauthswift: OAuth2Swift, _ oauthToken: String) {
         let _ = oauthswift.client.get(
-            "https://api.soundcloud.com/me?oauth_token=\(oauthToken)",
-            success: { response in
+        "https://api.soundcloud.com/me?oauth_token=\(oauthToken)") { result in
+            switch result {
+            case .success(let response):
                 let dataString = response.string!
                 print(dataString)
-            }, failure: { error in
+            case .failure(let error):
                 print(error)
             }
-        )
+        }
     }
     
     // MARK: Wordpress
@@ -1260,18 +1271,18 @@ extension ViewController {
         // The callback url you set here doesn't seem to make a differnce,
         // you have to set it up at the site when you get your developer key.
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/wordpress")!,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/wordpress")!) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-              
-            },
-            failure: { error in
+                
+            case .failure(let error):
                 print(error.localizedDescription, terminator: "")
             }
-        )
-
+        }
+        
     }
-
+    
     // MARK: Digu
     func doOAuthDigu(_ serviceParameters: [String:String]){
         let oauthswift = OAuth2Swift(
@@ -1285,16 +1296,16 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/digu")!, scope: "user,news,statuses", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/digu")!, scope: "user,news,statuses", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-            },
-                failure: { error in
-                    print(error.description)
+            case .failure(let error):
+                print(error.description)
             }
-        )
+        }
     }
-
+    
     // MARK: Noun
     func doOAuthNoun(_ serviceParameters: [String:String]) {
         let oauthswift = OAuth1Swift(
@@ -1308,17 +1319,18 @@ extension ViewController {
         
         self.testNoun(oauthswift)
     }
-
     func testNoun(_ oauthswift: OAuth1Swift) {
-        let _ = oauthswift.client.get("http://api.thenounproject.com/icon/apple",
-           success: { response in
-             let dataJSON = try? response.jsonObject()
-            print(String(describing: dataJSON))
-        }, failure: { error in
-            print(error)
-        })
+        let _ = oauthswift.client.get("http://api.thenounproject.com/icon/apple") { result in
+            switch result {
+            case .success(let response):
+                let dataJSON = try? response.jsonObject()
+                print(String(describing: dataJSON))
+            case .failure(let error):
+                print(error)
+            }
+        }
     }
-
+    
     // MARK: Lyft
     func doOAuthLyft(_ serviceParameters: [String:String]){
         let oauthswift = OAuth2Swift(
@@ -1333,14 +1345,94 @@ extension ViewController {
         oauthswift.authorizeURLHandler = getURLHandler()
         let state = generateState(withLength: 20)
         let _ = oauthswift.authorize(
-            withCallbackURL: URL(string: "oauth-swift://oauth-callback/lift")!, scope: "rides.read", state: state,
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/lift")!, scope: "rides.read", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
                 self.showTokenAlert(name: serviceParameters["name"], credential: credential)
-        },
-            failure: { error in
+            case .failure(let error):
                 print(error.description)
+            }
         }
+    }
+  
+    // MARK: Twitch
+    func doOAuthTwitch(_ serviceParameters: [String:String]){
+        let oauthswift = OAuth2Swift(
+            consumerKey:    serviceParameters["consumerKey"]!,
+            consumerSecret: serviceParameters["consumerSecret"]!,
+            authorizeUrl:   "https://id.twitch.tv/oauth2/authorize",
+            accessTokenUrl: "https://id.twitch.tv/oauth2/token",
+            responseType:   "code",
+            contentType:    "application/json"
         )
+        self.oauthswift = oauthswift
+        oauthswift.authorizeURLHandler = getURLHandler()
+        let state = generateState(withLength: 20)
+        let _ = oauthswift.authorize(
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/twitch")!, scope: "user_read", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
+                self.showTokenAlert(name: serviceParameters["name"], credential: credential)
+                self.testTwitch(oauthswift,credential.oauthToken)
+            case .failure(let error):
+                print(error.description)
+            }
+        }
+    }
+    
+    func testTwitch(_ oauthswift: OAuth2Swift, _ oauthToken: String) {
+        let _ = oauthswift.client.get(
+        "https://api.twitch.tv/kraken/user?oauth_token=\(oauthToken)", headers: ["Accept":"application/vnd.twitchtv.v5+json"]) { result in
+            switch result {
+            case .success(let response):
+                let dataString = response.string!
+                print(dataString)
+            case .failure(let error):
+                print(error)
+            }
+        }
+    }
+    
+    func doOauthReddit(_ serviceParameters: [String:String]) {
+        let oauthswift = OAuth2Swift(
+            consumerKey:    serviceParameters["consumerKey"]!,
+            consumerSecret: serviceParameters["consumerSecret"]!,
+            authorizeUrl:   "https://www.reddit.com/api/v1/authorize.compact",
+            accessTokenUrl: "https://www.reddit.com/api/v1/access_token",
+            responseType:   "code",
+            contentType:    "application/json"
+        )
+        self.oauthswift = oauthswift
+        oauthswift.accessTokenBasicAuthentification = true
+        oauthswift.authorizeURLHandler = getURLHandler()
+        let state = generateState(withLength: 20)
+        let _ = oauthswift.authorize(
+        withCallbackURL: URL(string: "oauth-swift://oauth-callback/reddit")!, scope: "read", state: state) { result in
+            switch result {
+            case .success(let (credential, _, _)):
+                self.showTokenAlert(name: serviceParameters["name"], credential: credential)
+                self.testReddit(oauthswift, credential.oauthToken)
+            case .failure(let error):
+                print(error.description)
+            }
+        }
+    }
+    
+    func testReddit(_ oauthswift: OAuth2Swift, _ oauthToken: String) {
+        let applicationName = ""    // Provided by developer
+        let username = ""           // Provided by developer
+        let userAgent = "ios:\(applicationName):1.0 (by /u/\(username)"
+        let headers = ["Authorization": "Bearer \(oauthToken)", "User-Agent": userAgent]
+        let _ = oauthswift.client.get(
+        "https://oauth.reddit.com/r/all", headers: headers) { result in
+            switch result {
+            case .success(let response):
+                let dataString = response.string!
+                print(dataString)
+            case .failure(let error):
+                print(error)
+            }
+        }
     }
 }
 
@@ -1360,11 +1452,11 @@ extension ViewController {
         let _ = internalWebViewController.webView
         
         #if os(iOS)
-            self.navigationItem.title = "OAuth"
-            let tableView: UITableView = UITableView(frame: self.view.bounds, style: .plain)
-            tableView.delegate = self
-            tableView.dataSource = self
-            self.view.addSubview(tableView)
+        self.navigationItem.title = "OAuth"
+        let tableView: UITableView = UITableView(frame: self.view.bounds, style: .plain)
+        tableView.delegate = self
+        tableView.dataSource = self
+        self.view.addSubview(tableView)
         #endif
     }
     
@@ -1399,6 +1491,9 @@ extension ViewController {
             }
         }
         services.loadFromFile(confPath)
+      
+        // Configure log level, if desired
+        OAuthSwift.setLogLevel(.trace)
     }
     
     func initConfOld() { // TODO Must be removed later
@@ -1427,30 +1522,30 @@ extension ViewController {
     
     func snapshot() -> Data {
         #if os(iOS)
-            UIGraphicsBeginImageContext(self.view.frame.size)
-            self.view.layer.render(in: UIGraphicsGetCurrentContext()!)
-            let fullScreenshot = UIGraphicsGetImageFromCurrentImageContext()
-            UIGraphicsEndImageContext()
-            UIImageWriteToSavedPhotosAlbum(fullScreenshot!, nil, nil, nil)
-            return fullScreenshot!.jpegData(compressionQuality: 0.5)!
+        UIGraphicsBeginImageContext(self.view.frame.size)
+        self.view.layer.render(in: UIGraphicsGetCurrentContext()!)
+        let fullScreenshot = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        UIImageWriteToSavedPhotosAlbum(fullScreenshot!, nil, nil, nil)
+        return fullScreenshot!.jpegData(compressionQuality: 0.5)!
         #elseif os(OSX)
-            let rep: NSBitmapImageRep = self.view.bitmapImageRepForCachingDisplay(in: self.view.bounds)!
-            self.view.cacheDisplay(in: self.view.bounds, to:rep)
-            return rep.tiffRepresentation!
+        let rep: NSBitmapImageRep = self.view.bitmapImageRepForCachingDisplay(in: self.view.bounds)!
+        self.view.cacheDisplay(in: self.view.bounds, to:rep)
+        return rep.tiffRepresentation!
         #endif
     }
     
     func showAlertView(title: String, message: String) {
         #if os(iOS)
-            let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertController.Style.alert)
-            alert.addAction(UIAlertAction(title: "Close", style: UIAlertAction.Style.default, handler: nil))
-            self.present(alert, animated: true, completion: nil)
+        let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertController.Style.alert)
+        alert.addAction(UIAlertAction(title: "Close", style: UIAlertAction.Style.default, handler: nil))
+        self.present(alert, animated: true, completion: nil)
         #elseif os(OSX)
-            let alert = NSAlert()
-            alert.messageText = title
-            alert.informativeText = message
-            alert.addButton(withTitle: "Close")
-            alert.runModal()
+        let alert = NSAlert()
+        alert.messageText = title
+        alert.informativeText = message
+        alert.addButton(withTitle: "Close")
+        alert.runModal()
         #endif
     }
     
@@ -1483,40 +1578,47 @@ extension ViewController {
             return internalWebViewController
         case .safari:
             #if os(iOS)
-                if #available(iOS 9.0, *) {
-                    let handler = SafariURLHandler(viewController: self, oauthSwift: self.oauthswift!)
-                    handler.presentCompletion = {
-                        print("Safari presented")
-                    }
-                    handler.dismissCompletion = {
-                        print("Safari dismissed")
-                    }
-                    handler.factory = { url in
-                        let controller = SFSafariViewController(url: url)
-                        // Customize it, for instance
-                        if #available(iOS 10.0, *) {
-                           //  controller.preferredBarTintColor = UIColor.red
-                        }
-                        return controller
-                    }
-                    
-                    return handler
+            if #available(iOS 9.0, *) {
+                let handler = SafariURLHandler(viewController: self, oauthSwift: self.oauthswift!)
+                handler.presentCompletion = {
+                    print("Safari presented")
                 }
+                handler.dismissCompletion = {
+                    print("Safari dismissed")
+                }
+                handler.factory = { url in
+                    let controller = SFSafariViewController(url: url)
+                    // Customize it, for instance
+                    if #available(iOS 10.0, *) {
+                        //  controller.preferredBarTintColor = UIColor.red
+                    }
+                    return controller
+                }
+                
+                return handler
+            }
+            #endif
+            return OAuthSwiftOpenURLExternally.sharedInstance
+        case .asWeb:
+            #if os(iOS)
+            if #available(iOS 13.0, tvOS 13.0, *) {
+                return ASWebAuthenticationURLHandler(callbackUrlScheme: "oauth-swift://oauth-callback/", presentationContextProvider: self)
+            }
             #endif
             return OAuthSwiftOpenURLExternally.sharedInstance
         }
-        
+
         #if os(OSX)
-            // a better way is
-            // - to make this ViewController implement OAuthSwiftURLHandlerType and assigned in oauthswift object
-            /* return self */
-            // - have an instance of WebViewController here (I) or a segue name to launch (S)
-            // - in handle(url)
-            //    (I) : affect url to WebViewController, and  self.presentViewControllerAsModalWindow(self.webViewController)
-            //    (S) : affect url to a temp variable (ex: urlForWebView), then perform segue
-            /* performSegueWithIdentifier("oauthwebview", sender:nil) */
-            //         then override prepareForSegue() to affect url to destination controller WebViewController
-            
+        // a better way is
+        // - to make this ViewController implement OAuthSwiftURLHandlerType and assigned in oauthswift object
+        /* return self */
+        // - have an instance of WebViewController here (I) or a segue name to launch (S)
+        // - in handle(url)
+        //    (I) : affect url to WebViewController, and  self.presentViewControllerAsModalWindow(self.webViewController)
+        //    (S) : affect url to a temp variable (ex: urlForWebView), then perform segue
+        /* performSegueWithIdentifier("oauthwebview", sender:nil) */
+        //         then override prepareForSegue() to affect url to destination controller WebViewController
+        
         #endif
     }
     //(I)
@@ -1528,9 +1630,9 @@ extension ViewController {
     override func prepare(for segue: OAuthStoryboardSegue, sender: Any?) {
         if segue.identifier == Storyboards.Main.formSegue {
             #if os(OSX)
-                let controller = segue.destinationController as? FormViewController
+            let controller = segue.destinationController as? FormViewController
             #else
-                let controller = segue.destination as? FormViewController
+            let controller = segue.destination as? FormViewController
             #endif
             // Fill the controller
             if let controller = controller {
@@ -1547,75 +1649,88 @@ public typealias Queue = DispatchQueue
 // MARK: - Table
 
 #if os(iOS)
-    extension ViewController: UITableViewDelegate, UITableViewDataSource {
-        // MARK: UITableViewDataSource
-        
-        func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-            return services.keys.count
-        }
-        func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-            let cell: UITableViewCell = UITableViewCell(style: UITableViewCell.CellStyle.subtitle, reuseIdentifier: "Cell")
-            let service = services.keys[indexPath.row]
-            cell.textLabel?.text = service
-            
-            if let parameters = services[service] , Services.parametersEmpty(parameters) {
-                cell.textLabel?.textColor = UIColor.red
-            }
-            if let parameters = services[service], let authentified = parameters["authentified"], authentified == "1" {
-                cell.textLabel?.textColor = UIColor.green
-            }
-            return cell
-        }
-        
-        // MARK: UITableViewDelegate
-          func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-            let service: String = services.keys[indexPath.row]
-            
-            DispatchQueue.global(qos: .background).async {
-                self.doAuthService(service: service)
-            }
-            tableView.deselectRow(at: indexPath, animated:true)
-        }
+extension ViewController: UITableViewDelegate, UITableViewDataSource {
+    // MARK: UITableViewDataSource
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return services.keys.count
     }
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell: UITableViewCell = UITableViewCell(style: UITableViewCell.CellStyle.subtitle, reuseIdentifier: "Cell")
+        let service = services.keys[indexPath.row]
+        cell.textLabel?.text = service
+        
+        if let parameters = services[service] , Services.parametersEmpty(parameters) {
+            cell.textLabel?.textColor = UIColor.red
+        }
+        if let parameters = services[service], let authentified = parameters["authentified"], authentified == "1" {
+            cell.textLabel?.textColor = UIColor.green
+        }
+        return cell
+    }
+    
+    // MARK: UITableViewDelegate
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let service: String = services.keys[indexPath.row]
+        
+        DispatchQueue.global(qos: .background).async {
+            self.doAuthService(service: service)
+        }
+        tableView.deselectRow(at: indexPath, animated:true)
+    }
+}
 #elseif os(OSX)
-    extension ViewController: NSTableViewDataSource, NSTableViewDelegate {
-        // MARK: NSTableViewDataSource
-        
-        func numberOfRows(in tableView: NSTableView) -> Int {
-            return services.keys.count
+extension ViewController: NSTableViewDataSource, NSTableViewDelegate {
+    // MARK: NSTableViewDataSource
+    
+    func numberOfRows(in tableView: NSTableView) -> Int {
+        return services.keys.count
+    }
+    
+    func tableView(_ tableView: NSTableView, objectValueFor tableColumn: NSTableColumn?, row: Int) -> Any? {
+        return services.keys[row]
+    }
+    
+    // MARK: NSTableViewDelegate
+    
+    func tableView(_ tableView: NSTableView, didAdd rowView: NSTableRowView, forRow row: Int) {
+        let service = services.keys[row]
+        if let parameters = services[service], Services.parametersEmpty(parameters) {
+            rowView.backgroundColor = NSColor.red
         }
-        
-        func tableView(_ tableView: NSTableView, objectValueFor tableColumn: NSTableColumn?, row: Int) -> Any? {
-            return services.keys[row]
+        if let parameters = services[service], let authentified = parameters["authentified"], authentified == "1" {
+            rowView.backgroundColor  = NSColor.green
         }
-        
-        // MARK: NSTableViewDelegate
-        
-        func tableView(_ tableView: NSTableView, didAdd rowView: NSTableRowView, forRow row: Int) {
-            let service = services.keys[row]
-            if let parameters = services[service], Services.parametersEmpty(parameters) {
-                rowView.backgroundColor = NSColor.red
-            }
-            if let parameters = services[service], let authentified = parameters["authentified"], authentified == "1" {
-                rowView.backgroundColor  = NSColor.green
-            }
-        }
-        
-        func tableViewSelectionDidChange(_ notification: Notification) {
-            if let tableView = notification.object as? NSTableView {
-                let row = tableView.selectedRow
-                if  row != -1 {
-                    let service: String = services.keys[row]
-                    
-                    
-                    DispatchQueue.global(qos: .background).async {
-                        self.doAuthService(service: service)
-                    }
-                    tableView.deselectRow(row)
+    }
+    
+    func tableViewSelectionDidChange(_ notification: Notification) {
+        if let tableView = notification.object as? NSTableView {
+            let row = tableView.selectedRow
+            if  row != -1 {
+                let service: String = services.keys[row]
+                
+                
+                DispatchQueue.global(qos: .background).async {
+                    self.doAuthService(service: service)
                 }
+                tableView.deselectRow(row)
             }
         }
     }
+}
+#endif
+
+#if os(iOS)
+import SafariServices
+#if canImport(AuthenticationServices)
+import AuthenticationServices
+#endif
+@available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *)
+extension ViewController: ASWebAuthenticationPresentationContextProviding {
+    func presentationAnchor(for session: ASWebAuthenticationSession) -> ASPresentationAnchor {
+        return UIApplication.shared.topWindow ?? ASPresentationAnchor()
+    }
+}
 #endif
 
 struct FormViewControllerData {
@@ -1643,10 +1758,10 @@ extension ViewController: FormViewControllerDelegate {
     
     func dismissForm() {
         #if os(iOS)
-            /*self.dismissViewControllerAnimated(true) { // without animation controller
-             print("form dismissed")
-             }*/
-            let _ = self.navigationController?.popViewController(animated: true)
+        /*self.dismissViewControllerAnimated(true) { // without animation controller
+         print("form dismissed")
+         }*/
+        let _ = self.navigationController?.popViewController(animated: true)
         #endif
     }
 }
