@@ -12,6 +12,10 @@ import XCTest
 class OAuth2SwiftTests: XCTestCase {
     
     static let server = TestServer()
+    
+    //Using NSLock for Linux compatible locking
+    let serverLock = NSLock()
+    
     var server: TestServer { return OAuth2SwiftTests.server }
 
     override class func setUp() {
@@ -32,21 +36,21 @@ class OAuth2SwiftTests: XCTestCase {
     let callbackURL = "test://callback"
 
     func testDataSuccess() {
-        objc_sync_enter(server)
+       serverLock.lock()
         let state = generateState(withLength: 20)
         testSuccess(.data, response: .code("code", state:state))
-        objc_sync_exit(server)
+        serverLock.unlock()
     }
     func testJSON_Code_Success() {
-        objc_sync_enter(server)
+       serverLock.lock()
         let state = generateState(withLength: 20)
         testSuccess(.json, response: .code("code", state:state))
-        objc_sync_exit(server)
+        serverLock.unlock()
     }
     func testJSON_AccessToken_Success() {
-        objc_sync_enter(server)
+       serverLock.lock()
         testSuccess(.json, response: .accessToken(server.oauth_token))
-        objc_sync_exit(server)
+        serverLock.unlock()
     }
 
     func testSuccess(_ accessReturnType: TestServer.AccessReturnType, response: AccessTokenResponse) {
@@ -74,14 +78,14 @@ class OAuth2SwiftTests: XCTestCase {
 			state = extractedState ?? ""
 		}
         let _ = oauth.authorize(
-            withCallbackURL: URL(string:callbackURL)!, scope: "all", state: state, parameters: [:],
-            success: { credential, response, parameters in
+            withCallbackURL: URL(string:callbackURL)!, scope: "all", state: state, parameters: [:]) { result in
+            switch result {
+            case .success:
                 expectation.fulfill()
-            },
-            failure: { error in
+            case .failure(let error):
                 XCTFail("The failure handler should not be called.\(error)")
             }
-        )
+        }
         
         waitForExpectations(timeout: DefaultTimeout, handler: nil)
         
@@ -89,15 +93,15 @@ class OAuth2SwiftTests: XCTestCase {
     }
     
     func testJSON_Error_Failure() {
-        objc_sync_enter(server)
+       serverLock.lock()
         testFailure(.json, response: .error("bad", "very bad"))
-        objc_sync_exit(server)
+        serverLock.unlock()
     }
 
     func testJSON_None_Failure() {
-        objc_sync_enter(server)
+       serverLock.lock()
         testFailure(.json, response: .none)
-        objc_sync_exit(server)
+        serverLock.unlock()
     }
     
     func testFailure(_ accessReturnType: TestServer.AccessReturnType, response: AccessTokenResponse) {
@@ -122,22 +126,21 @@ class OAuth2SwiftTests: XCTestCase {
         
         let state = generateState(withLength: 20)
         let _ = oauth.authorize(
-            withCallbackURL: URL(string:callbackURL)!, scope: "all", state: state, parameters: [:],
-            success: { credential, response, parameters in
+        withCallbackURL: URL(string:callbackURL)!, scope: "all", state: state, parameters: [:]) { result in
+            switch result {
+            case .success:
                 XCTFail("The success handler should not be called.")
-            },
-            failure: { error in
+            case .failure:
                 expectation.fulfill()
             }
-        )
+        }
         
         waitForExpectations(timeout: DefaultTimeout, handler: nil)
     }
     
     func testExpire() {
         let expectation = self.expectation(description: "request should failed")
-        
-        
+
         let oauth = OAuth2Swift(
             consumerKey: server.valid_key,
             consumerSecret: server.valid_secret,
@@ -145,12 +148,11 @@ class OAuth2SwiftTests: XCTestCase {
             accessTokenUrl: server.accessTokenURLV2,
             responseType: "code"
         )
-        let _ = oauth.client.get(
-            server.expireURLV2, parameters: [:],
-            success: { response in
+        let _ = oauth.client.get(server.expireURLV2, parameters: [:]) { result in
+            switch result {
+            case .success(let response):
                 XCTFail("data receive \(response.data).")
-            },
-            failure: { error in
+            case .failure(let error):
                 switch error {
                 case .tokenExpired(let error):
                     expectation.fulfill()
@@ -175,9 +177,7 @@ class OAuth2SwiftTests: XCTestCase {
                     XCTFail("Wrong exception type \(error)")
                 }
             }
-        )
+        }
         waitForExpectations(timeout: DefaultTimeout, handler: nil)
     }
-    
-
 }
