@@ -18,15 +18,15 @@ public extension NSError {
     ///
     /// Also implements a special handling for the Facebook API, which indicates invalid tokens in a 
     /// different manner. See https://developers.facebook.com/docs/graph-api/using-graph-api#errors
-	public var isExpiredToken: Bool {
-        guard self.domain == NSURLErrorDomain else {
+    var isExpiredToken: Bool {
+        guard self.domain == NSURLErrorDomain || self.domain == OAuthSwiftError.Domain else {
             return false
         }
 		if self.code == 401 {
 			if let reponseHeaders = self.userInfo["Response-Headers"] as? [String: String],
 				let authenticateHeader = reponseHeaders["WWW-Authenticate"] ?? reponseHeaders["Www-Authenticate"] {
 				let headerDictionary = authenticateHeader.headerDictionary
-				if let error = headerDictionary["error"], error == "invalid_token" || error == "\"invalid_token\"" {
+				if let error = headerDictionary["error"], error == "invalid_token" || error == "expired_token" || error == "\"invalid_token\"" {
 					return true
 				}
 			}
@@ -34,12 +34,16 @@ public extension NSError {
                 let bodyData = body.data(using: OAuthSwiftDataEncoding),
                 let json = try? JSONSerialization.jsonObject(with: bodyData, options: []),
                 let jsonDic = json as? [String: AnyObject] {
-                if let error = jsonDic["error"] as? String, error == "invalid_token" || error == "\"invalid_token\"" {
+                if let error = jsonDic["error"] as? String, error == "invalid_token" || error == "expired_token" || error == "\"invalid_token\"" {
                     return true
                 }
                 if let errors = jsonDic["errors"] as? [[String: AnyObject]] {
                     for error in errors {
-                        if let errorType = error["errorType"] as? String, errorType == "invalid_token" {
+                        if let errorType = error["errorType"] as? String, errorType == "invalid_token" || errorType == "expired_token" {
+                            return true
+                        } else if let urlString = self.userInfo[NSURLErrorFailingURLErrorKey] as? String, urlString.contains("api.twitter.com"), let errorCode = error["code"] as? Int, errorCode == 89 {
+                            // This is the code for expired or invalid twitter token
+                            // see: https://developer.twitter.com/en/docs/basics/response-codes
                             return true
                         }
                     }
