@@ -30,23 +30,16 @@ open class OAuth2Swift: OAuthSwift {
     var accessTokenUrl: String?
     var responseType: String
     var contentType: String?
+    // RFC7636 PKCE
+    var codeVerifier: String?
 
     // MARK: init
-    public convenience init(consumerKey: String, consumerSecret: String, authorizeUrl: String, accessTokenUrl: String, responseType: String) {
-        self.init(consumerKey: consumerKey, consumerSecret: consumerSecret, authorizeUrl: authorizeUrl, responseType: responseType)
-        self.accessTokenUrl = accessTokenUrl
-    }
-
-    public convenience init(consumerKey: String, consumerSecret: String, authorizeUrl: String, accessTokenUrl: String, responseType: String, contentType: String) {
-        self.init(consumerKey: consumerKey, consumerSecret: consumerSecret, authorizeUrl: authorizeUrl, responseType: responseType)
-        self.accessTokenUrl = accessTokenUrl
+    public init(consumerKey: String, consumerSecret: String, authorizeUrl: URLConvertible, accessTokenUrl: URLConvertible? = nil, responseType: String, contentType: String? = nil) {
+        self.accessTokenUrl = accessTokenUrl?.string
         self.contentType = contentType
-    }
-
-    public init(consumerKey: String, consumerSecret: String, authorizeUrl: String, responseType: String) {
         self.consumerKey = consumerKey
         self.consumerSecret = consumerSecret
-        self.authorizeUrl = authorizeUrl
+        self.authorizeUrl = authorizeUrl.string
         self.responseType = responseType
         super.init(consumerKey: consumerKey, consumerSecret: consumerSecret)
         self.client.credential.version = .oauth2
@@ -54,15 +47,15 @@ open class OAuth2Swift: OAuthSwift {
 
     public convenience init?(parameters: ConfigParameters) {
         guard let consumerKey = parameters["consumerKey"], let consumerSecret = parameters["consumerSecret"],
-              let responseType = parameters["responseType"], let authorizeUrl = parameters["authorizeUrl"] else {
-            return nil
+            let responseType = parameters["responseType"], let authorizeUrl = parameters["authorizeUrl"] else {
+                return nil
         }
         if let accessTokenUrl = parameters["accessTokenUrl"] {
             self.init(consumerKey: consumerKey, consumerSecret: consumerSecret,
-                authorizeUrl: authorizeUrl, accessTokenUrl: accessTokenUrl, responseType: responseType)
+                      authorizeUrl: authorizeUrl, accessTokenUrl: accessTokenUrl, responseType: responseType)
         } else {
             self.init(consumerKey: consumerKey, consumerSecret: consumerSecret,
-                authorizeUrl: authorizeUrl, responseType: responseType)
+                      authorizeUrl: authorizeUrl, responseType: responseType)
         }
     }
 
@@ -78,11 +71,18 @@ open class OAuth2Swift: OAuthSwift {
 
     // MARK: functions
     @discardableResult
-    open func authorize(withCallbackURL callbackURL: URL?, scope: String, state: String, parameters: Parameters = [:], headers: OAuthSwift.Headers? = nil, success: @escaping TokenSuccessHandler, failure: FailureHandler?) -> OAuthSwiftRequestHandle? {
+    open func authorize(withCallbackURL callbackURL: URLConvertible?, scope: String, state: String, parameters: Parameters = [:], headers: OAuthSwift.Headers? = nil, completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
 
+        OAuthSwift.log?.trace("Start authorization ...")
+        if let url = callbackURL, url.url == nil {
+            completion(.failure(.encodingError(urlString: url.string)))
+            return nil
+        }
         self.observeCallback { [weak self] url in
+
+            OAuthSwift.log?.trace("Open application resource url: \(url.absoluteString)")
             guard let this = self else {
-                OAuthSwift.retainError(failure)
+                OAuthSwift.retainError(completion)
                 return
             }
             var responseParameters = [String: String]()
@@ -92,41 +92,81 @@ open class OAuth2Swift: OAuthSwift {
             if let fragment = url.fragment, !fragment.isEmpty {
                 responseParameters += fragment.parametersFromQueryString
             }
+            OAuthSwift.log?.trace("Parsed url parameters: \(responseParameters)")
+
             if let accessToken = responseParameters["access_token"] {
                 this.client.credential.oauthToken = accessToken.safeStringByRemovingPercentEncoding
                 if let expiresIn: String = responseParameters["expires_in"], let offset = Double(expiresIn) {
                     this.client.credential.oauthTokenExpiresAt = Date(timeInterval: offset, since: Date())
                 }
-                success(this.client.credential, nil, responseParameters)
+                completion(.success((this.client.credential, nil, responseParameters)))
             } else if let code = responseParameters["code"] {
                 if !this.allowMissingStateCheck {
                     guard let responseState = responseParameters["state"] else {
-                        failure?(OAuthSwiftError.missingState)
+                        OAuthSwift.log?.error("Resource url: Missing 'state' parameter")
+                        completion(.failure(.missingState))
                         return
                     }
                     if responseState != state {
-                        failure?(OAuthSwiftError.stateNotEqual(state: state, responseState: responseState))
+                        OAuthSwift.log?.error("Resource url: Unmatched 'state' parameter")
+                        completion(.failure(.stateNotEqual(state: state, responseState: responseState)))
                         return
                     }
                 }
                 let callbackURLEncoded: URL?
                 if let callbackURL = callbackURL {
-                    callbackURLEncoded = URL(string: callbackURL.absoluteString.urlEncoded)!
+                    callbackURLEncoded = callbackURL.encodedURL // XXX do not known why to re-encode, maybe if string only?
                 } else {
                     callbackURLEncoded = nil
                 }
                 if let handle = this.postOAuthAccessTokenWithRequestToken(
                     byCode: code.safeStringByRemovingPercentEncoding,
-                    callbackURL: callbackURLEncoded, headers: headers, success: success, failure: failure) {
+                    callbackURL: callbackURLEncoded, headers: headers, completionHandler: completion) {
                     this.putHandle(handle, withKey: UUID().uuidString)
                 }
             } else if let error = responseParameters["error"] {
-                let description = responseParameters["error_description"] ?? ""
-                let message = NSLocalizedString(error, comment: description)
-                failure?(OAuthSwiftError.serverError(message: message))
+                let otherErrorBlock = {
+                    let description = responseParameters["error_description"] ?? ""
+                    let message = NSLocalizedString(error, comment: description)
+                    OAuthSwift.log?.error("Authorization failed with: \(description)")
+                    completion(.failure(.serverError(message: message)))
+                }
+
+                // handling SFAuthenticationSession/ASWebAuthenticationSession canceledLogin errors
+                if let domain = responseParameters["error_domain"],
+                    let codeString = responseParameters["error_code"],
+                    let code = Int(codeString) {
+
+#if targetEnvironment(macCatalyst)
+					if #available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *),
+                        ASWebAuthenticationURLHandler.isCancelledError(domain: domain, code: code) {
+                        completion(.failure(.cancelled))
+                    } else {
+                        otherErrorBlock()
+                    }
+#else
+
+#if os(macOS)
+                    otherErrorBlock()
+#else
+                    if #available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *),
+                        ASWebAuthenticationURLHandler.isCancelledError(domain: domain, code: code) {
+                        completion(.failure(.cancelled))
+                    } else if #available(iOS 11, *),
+                        SFAuthenticationURLHandler.isCancelledError(domain: domain, code: code) {
+                        completion(.failure(.cancelled))
+                    } else {
+                        otherErrorBlock()
+                    }
+#endif
+#endif
+                } else {
+                    otherErrorBlock()
+                }
             } else {
                 let message = "No access_token, no code and no error provided by server"
-                failure?(OAuthSwiftError.serverError(message: message))
+                OAuthSwift.log?.error("Authorization failed with: \(message)")
+                completion(.failure(.serverError(message: message)))
             }
         }
 
@@ -140,7 +180,8 @@ open class OAuth2Swift: OAuthSwift {
         var queryString: String? = ""
         queryString = queryString?.urlQueryByAppending(parameter: "client_id", value: self.consumerKey, encodeError)
         if let callbackURL = callbackURL {
-            queryString = queryString?.urlQueryByAppending(parameter: "redirect_uri", value: self.encodeCallbackURL ? callbackURL.absoluteString.urlEncoded : callbackURL.absoluteString, encode: self.encodeCallbackURLQuery, encodeError)
+            let value = self.encodeCallbackURL ? callbackURL.string.urlEncoded : callbackURL.string
+            queryString = queryString?.urlQueryByAppending(parameter: "redirect_uri", value: value, encode: self.encodeCallbackURLQuery, encodeError)
         }
         queryString = queryString?.urlQueryByAppending(parameter: "response_type", value: self.responseType, encodeError)
         queryString = queryString?.urlQueryByAppending(parameter: "scope", value: scope, encodeError)
@@ -156,108 +197,52 @@ open class OAuth2Swift: OAuthSwift {
                 self.authorizeURLHandler.handle(url)
                 return self
             } else {
-                failure?(OAuthSwiftError.encodingError(urlString: urlString))
+                OAuthSwift.log?.error("Resource url: Invalid query string: \(urlString)")
+                completion(.failure(.encodingError(urlString: urlString)))
             }
         } else {
             let urlString = self.authorizeUrl.urlByAppending(query: queryErrorString)
-            failure?(OAuthSwiftError.encodingError(urlString: urlString))
+            OAuthSwift.log?.error("Resource url: Invalid query string: \(urlString)")
+            completion(.failure(.encodingError(urlString: urlString)))
         }
         self.cancel() // ie. remove the observer.
         return nil
     }
 
-    @discardableResult
-    open func authorize(withCallbackURL urlString: String, scope: String, state: String, parameters: Parameters = [:], headers: OAuthSwift.Headers? = nil, success: @escaping TokenSuccessHandler, failure: FailureHandler?) -> OAuthSwiftRequestHandle? {
-        guard let url = URL(string: urlString) else {
-            failure?(OAuthSwiftError.encodingError(urlString: urlString))
-            return nil
-        }
-        return authorize(withCallbackURL: url, scope: scope, state: state, parameters: parameters, headers: headers, success: success, failure: failure)
-    }
+    open func postOAuthAccessTokenWithRequestToken(byCode code: String, callbackURL: URL?, headers: OAuthSwift.Headers? = nil, completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
 
-    open func postOAuthAccessTokenWithRequestToken(byCode code: String, callbackURL: URL?, headers: OAuthSwift.Headers? = nil, success: @escaping TokenSuccessHandler, failure: FailureHandler?) -> OAuthSwiftRequestHandle? {
         var parameters = OAuthSwift.Parameters()
         parameters["client_id"] = self.consumerKey
-        parameters["client_secret"] = self.consumerSecret
         parameters["code"] = code
         parameters["grant_type"] = "authorization_code"
+
+        // PKCE - extra parameter
+        if let codeVerifier = self.codeVerifier {
+            parameters["code_verifier"] = codeVerifier
+            // Don't send client secret when using PKCE, some services complain
+        } else {
+            // client secrets should only be used for web style apps where they can't be decompiled (use pkce instead), so if it's empty, don't post it as some servers will reject it
+            // https://www.oauth.com/oauth2-servers/client-registration/client-id-secret/
+            if !self.consumerSecret.isEmpty {
+                parameters["client_secret"] = self.consumerSecret
+            }
+        }
+
         if let callbackURL = callbackURL {
             parameters["redirect_uri"] = callbackURL.absoluteString.safeStringByRemovingPercentEncoding
         }
 
-        return requestOAuthAccessToken(withParameters: parameters, headers: headers, success: success, failure: failure)
+        OAuthSwift.log?.trace("Add security parameters: \(parameters)")
+        return requestOAuthAccessToken(withParameters: parameters, headers: headers, completionHandler: completion)
     }
 
     @discardableResult
-    open func renewAccessToken(withRefreshToken refreshToken: String, parameters: OAuthSwift.Parameters? = nil, headers: OAuthSwift.Headers? = nil, success: @escaping TokenSuccessHandler, failure: FailureHandler?) -> OAuthSwiftRequestHandle? {
-        var parameters = parameters ?? OAuthSwift.Parameters()
-        parameters["client_id"] = self.consumerKey
-        parameters["client_secret"] = self.consumerSecret
-        parameters["refresh_token"] = refreshToken
-        parameters["grant_type"] = "refresh_token"
-
-        return requestOAuthAccessToken(withParameters: parameters, headers: headers, success: success, failure: failure)
+    open func renewAccessToken(withRefreshToken refreshToken: String, parameters: OAuthSwift.Parameters? = nil, headers: OAuthSwift.Headers? = nil, completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
+        return self.client.renewAccessToken(accessTokenUrl: self.accessTokenUrl, withRefreshToken: refreshToken, parameters: parameters ?? OAuthSwift.Parameters(), headers: headers, completionHandler: completion)
     }
 
-    fileprivate func requestOAuthAccessToken(withParameters parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, success: @escaping TokenSuccessHandler, failure: FailureHandler?) -> OAuthSwiftRequestHandle? {
-        let successHandler: OAuthSwiftHTTPRequest.SuccessHandler = { [weak self] response in
-            guard let this = self else {
-                OAuthSwift.retainError(failure)
-                return
-            }
-            let responseJSON: Any? = try? response.jsonObject(options: .mutableContainers)
-
-            let responseParameters: OAuthSwift.Parameters
-
-            if let jsonDico = responseJSON as? [String: Any] {
-                responseParameters = jsonDico
-            } else {
-                responseParameters = response.string?.parametersFromQueryString ?? [:]
-            }
-
-            guard let accessToken = responseParameters["access_token"] as? String else {
-                let message = NSLocalizedString("Could not get Access Token", comment: "Due to an error in the OAuth2 process, we couldn't get a valid token.")
-                failure?(OAuthSwiftError.serverError(message: message))
-                return
-            }
-
-            if let refreshToken = responseParameters["refresh_token"] as? String {
-                this.client.credential.oauthRefreshToken = refreshToken.safeStringByRemovingPercentEncoding
-            }
-
-            if let expiresIn = responseParameters["expires_in"] as? String, let offset = Double(expiresIn) {
-                this.client.credential.oauthTokenExpiresAt = Date(timeInterval: offset, since: Date())
-            } else if let expiresIn = responseParameters["expires_in"] as? Double {
-                this.client.credential.oauthTokenExpiresAt = Date(timeInterval: expiresIn, since: Date())
-            }
-
-            this.client.credential.oauthToken = accessToken.safeStringByRemovingPercentEncoding
-            success(this.client.credential, response, responseParameters)
-        }
-
-        guard let accessTokenUrl = accessTokenUrl else {
-            let message = NSLocalizedString("access token url not defined", comment: "access token url not defined with code type auth")
-            failure?(OAuthSwiftError.configurationError(message: message))
-            return nil
-        }
-
-        if self.contentType == "multipart/form-data" {
-            // Request new access token by disabling check on current token expiration. This is safe because the implementation wants the user to retrieve a new token.
-            return self.client.postMultiPartRequest(accessTokenUrl, method: .POST, parameters: parameters, headers: headers, checkTokenExpiration: false, success: successHandler, failure: failure)
-        } else {
-            // special headers
-            var finalHeaders: OAuthSwift.Headers? = headers
-            if accessTokenBasicAuthentification {
-
-                let authentification = "\(self.consumerKey):\(self.consumerSecret)".data(using: String.Encoding.utf8)
-                if let base64Encoded = authentification?.base64EncodedString(options: Data.Base64EncodingOptions(rawValue: 0)) {
-                    finalHeaders += ["Authorization": "Basic \(base64Encoded)"] as OAuthSwift.Headers
-                }
-            }
-
-            // Request new access token by disabling check on current token expiration. This is safe because the implementation wants the user to retrieve a new token.
-            return self.client.request(accessTokenUrl, method: .POST, parameters: parameters, headers: finalHeaders, checkTokenExpiration: false, success: successHandler, failure: failure)
-        }
+    fileprivate func requestOAuthAccessToken(withParameters parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
+        return self.client.requestOAuthAccessToken(accessTokenUrl: self.accessTokenUrl, withParameters: parameters, headers: headers, contentType: self.contentType, accessTokenBasicAuthentification: self.accessTokenBasicAuthentification, completionHandler: completion)
     }
 
     /**
@@ -276,37 +261,54 @@ open class OAuth2Swift: OAuthSwift {
      - parameter failure:        The failure block. Takes the error as parameter.
      */
     @discardableResult
-    open func startAuthorizedRequest(_ url: String, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, renewHeaders: OAuthSwift.Headers? = nil, body: Data? = nil, onTokenRenewal: TokenRenewedHandler? = nil, success: @escaping OAuthSwiftHTTPRequest.SuccessHandler, failure: @escaping OAuthSwiftHTTPRequest.FailureHandler) -> OAuthSwiftRequestHandle? {
-        // build request
-        return self.client.request(url, method: method, parameters: parameters, headers: headers, body: body, success: success) { (error) in
-            switch error {
+    open func startAuthorizedRequest(_ url: URLConvertible, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, renewHeaders: OAuthSwift.Headers? = nil, body: Data? = nil, onTokenRenewal: TokenRenewedHandler? = nil, completionHandler completion: @escaping OAuthSwiftHTTPRequest.CompletionHandler) -> OAuthSwiftRequestHandle? {
 
-            case OAuthSwiftError.tokenExpired:
-                _ = self.renewAccessToken(withRefreshToken: self.client.credential.oauthRefreshToken, headers: renewHeaders ?? headers, success: { (credential, _, _) in
-                    // Ommit response parameters so they don't override the original ones
-                    // We have successfully renewed the access token.
+        OAuthSwift.log?.trace("Start authorized request, url: \(url.url?.absoluteString ?? "unknown") ...")
+        let completionHandler: OAuthSwiftHTTPRequest.CompletionHandler = { result in
+            switch result {
+            case .success:
+                completion(result)
+            case .failure(let error): // map/recovery error
+                switch error {
+                case OAuthSwiftError.tokenExpired:
+                    let renewCompletionHandler: TokenCompletionHandler = { result in
+                        switch result {
+                        case .success(let (credential, _, _)):
+                            // Ommit response parameters so they don't override the original ones
+                            // We have successfully renewed the access token.
 
-                    // If provided, fire the onRenewal closure
-                    if let renewalCallBack = onTokenRenewal {
-                        renewalCallBack(credential)
+                            // If provided, fire the onRenewal closure
+                            if let renewalCallBack = onTokenRenewal {
+                                renewalCallBack(.success(credential))
+                            }
+
+                            // Reauthorize the request again, this time with a brand new access token ready to be used.
+                            _ = self.startAuthorizedRequest(url, method: method, parameters: parameters, headers: headers, body: body, onTokenRenewal: onTokenRenewal, completionHandler: completion)
+                        case .failure(let error):
+                            completion(.failure(error))
+                        }
                     }
 
-                    // Reauthorize the request again, this time with a brand new access token ready to be used.
-                    _ = self.startAuthorizedRequest(url, method: method, parameters: parameters, headers: headers, body: body, onTokenRenewal: onTokenRenewal, success: success, failure: failure)
-                }, failure: failure)
-            default:
-                failure(error)
+                    _ = self.renewAccessToken(withRefreshToken: self.client.credential.oauthRefreshToken, headers: renewHeaders ?? headers, completionHandler: renewCompletionHandler)
+                default:
+                    completion(.failure(error))
+                }
             }
         }
+
+        // build request
+        return self.client.request(url, method: method, parameters: parameters, headers: headers, body: body, completionHandler: completionHandler)
     }
 
-	// OAuth 2.0 Specification: https://tools.ietf.org/html/draft-ietf-oauth-v2-13#section-4.3
+    // OAuth 2.0 Specification: https://tools.ietf.org/html/draft-ietf-oauth-v2-13#section-4.3
     @discardableResult
-    open func authorize(username: String, password: String, scope: String?, headers: OAuthSwift.Headers? = nil, success: @escaping TokenSuccessHandler, failure: @escaping OAuthSwiftHTTPRequest.FailureHandler) -> OAuthSwiftRequestHandle? {
+    open func authorize(username: String, password: String, scope: String?, headers: OAuthSwift.Headers? = nil, completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
 
         var parameters = OAuthSwift.Parameters()
         parameters["client_id"] = self.consumerKey
-        parameters["client_secret"] = self.consumerSecret
+        if !self.consumerSecret.isEmpty {
+            parameters["client_secret"] = self.consumerSecret
+        }
         parameters["username"] = username
         parameters["password"] = password
         parameters["grant_type"] = "password"
@@ -318,13 +320,12 @@ open class OAuth2Swift: OAuthSwift {
         return requestOAuthAccessToken(
             withParameters: parameters,
             headers: headers,
-            success: success,
-            failure: failure
+            completionHandler: completion
         )
     }
 
     @discardableResult
-    open func authorize(deviceToken deviceCode: String, grantType: String = "http://oauth.net/grant_type/device/1.0", success: @escaping TokenSuccessHandler, failure: @escaping OAuthSwiftHTTPRequest.FailureHandler) -> OAuthSwiftRequestHandle? {
+    open func authorize(deviceToken deviceCode: String, grantType: String = "http://oauth.net/grant_type/device/1.0", completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
         var parameters = OAuthSwift.Parameters()
         parameters["client_id"] = self.consumerKey
         parameters["client_secret"] = self.consumerSecret
@@ -333,9 +334,25 @@ open class OAuth2Swift: OAuthSwift {
 
         return requestOAuthAccessToken(
             withParameters: parameters,
-            success: success,
-            failure: failure
+            completionHandler: completion
         )
     }
 
+    /// use RFC7636 PKCE credentials - convenience method
+    @discardableResult
+    open func authorize(withCallbackURL url: URLConvertible, scope: String, state: String, codeChallenge: String, codeChallengeMethod: String = "S256", codeVerifier: String, parameters: Parameters = [:], headers: OAuthSwift.Headers? = nil, completionHandler completion: @escaping TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
+        guard let callbackURL = url.url else {
+            completion(.failure(.encodingError(urlString: url.string)))
+            return nil
+        }
+
+        // remember code_verifier
+        self.codeVerifier = codeVerifier
+        // PKCE - extra parameter
+        var pkceParameters = Parameters()
+        pkceParameters["code_challenge"] = codeChallenge
+        pkceParameters["code_challenge_method"] = codeChallengeMethod
+
+        return authorize(withCallbackURL: callbackURL, scope: scope, state: state, parameters: parameters + pkceParameters, headers: headers, completionHandler: completion)
+    }
 }
