@@ -45,45 +45,45 @@ open class OAuthSwiftClient: NSObject {
 
     // MARK: client methods
     @discardableResult
-    open func get(_ urlString: String, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
-        return self.request(urlString, method: .GET, parameters: parameters, headers: headers, success: success, failure: failure)
+    open func get(_ url: URLConvertible, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
+        return self.request(url, method: .GET, parameters: parameters, headers: headers, completionHandler: completion)
     }
 
     @discardableResult
-    open func post(_ urlString: String, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
-        return self.request(urlString, method: .POST, parameters: parameters, headers: headers, body: body, success: success, failure: failure)
+    open func post(_ url: URLConvertible, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
+        return self.request(url, method: .POST, parameters: parameters, headers: headers, body: body, completionHandler: completion)
     }
 
     @discardableResult
-    open func put(_ urlString: String, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
-        return self.request(urlString, method: .PUT, parameters: parameters, headers: headers, body: body, success: success, failure: failure)
+    open func put(_ url: URLConvertible, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
+        return self.request(url, method: .PUT, parameters: parameters, headers: headers, body: body, completionHandler: completion)
     }
 
     @discardableResult
-    open func delete(_ urlString: String, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
-        return self.request(urlString, method: .DELETE, parameters: parameters, headers: headers, success: success, failure: failure)
+    open func delete(_ url: URLConvertible, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
+        return self.request(url, method: .DELETE, parameters: parameters, headers: headers, completionHandler: completion)
     }
 
     @discardableResult
-    open func patch(_ urlString: String, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
-        return self.request(urlString, method: .PATCH, parameters: parameters, headers: headers, success: success, failure: failure)
+    open func patch(_ url: URLConvertible, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
+        return self.request(url, method: .PATCH, parameters: parameters, headers: headers, completionHandler: completion)
     }
 
     @discardableResult
-    open func request(_ urlString: String, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil, checkTokenExpiration: Bool = true, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
+    open func request(_ url: URLConvertible, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil, checkTokenExpiration: Bool = true, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
 
         if checkTokenExpiration && self.credential.isTokenExpired() {
-            failure?(OAuthSwiftError.tokenExpired(error: nil))
+            completion?(.failure(.tokenExpired(error: nil)))
             return nil
         }
 
-        guard URL(string: urlString) != nil else {
-            failure?(OAuthSwiftError.encodingError(urlString: urlString))
+        guard url.url != nil else {
+            completion?(.failure(.encodingError(urlString: url.string)))
             return nil
         }
 
-        if let request = makeRequest(urlString, method: method, parameters: parameters, headers: headers, body: body) {
-            request.start(success: success, failure: failure)
+        if let request = makeRequest(url, method: method, parameters: parameters, headers: headers, body: body) {
+            request.start(completionHandler: completion)
             return request
         }
         return nil
@@ -95,9 +95,9 @@ open class OAuthSwiftClient: NSObject {
         return request
     }
 
-    open func makeRequest(_ urlString: String, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil) -> OAuthSwiftHTTPRequest? {
-        guard let url = URL(string: urlString) else {
-            return nil
+    open func makeRequest(_ url: URLConvertible, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, body: Data? = nil) -> OAuthSwiftHTTPRequest? {
+        guard let url = url.url else {
+            return nil // XXX failure not thrown here
         }
 
         let request = OAuthSwiftHTTPRequest(url: url, method: method, parameters: parameters, paramsLocation: self.paramsLocation, httpBody: body, headers: headers ?? [:], sessionFactory: self.sessionFactory)
@@ -106,11 +106,11 @@ open class OAuthSwiftClient: NSObject {
     }
 
     @discardableResult
-    public func postImage(_ urlString: String, parameters: OAuthSwift.Parameters, image: Data, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
-        return self.multiPartRequest(url: urlString, method: .POST, parameters: parameters, image: image, success: success, failure: failure)
+    public func postImage(_ url: URLConvertible, parameters: OAuthSwift.Parameters, image: Data, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
+        return self.multiPartRequest(url: url, method: .POST, parameters: parameters, image: image, completionHandler: completion)
     }
 
-    open func makeMultiPartRequest(_ urlString: String, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], multiparts: [OAuthSwiftMultipartData] = [], headers: OAuthSwift.Headers? = nil) -> OAuthSwiftHTTPRequest? {
+    open func makeMultiPartRequest(_ url: URLConvertible, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], multiparts: [OAuthSwiftMultipartData] = [], headers: OAuthSwift.Headers? = nil) -> OAuthSwiftHTTPRequest? {
         let boundary = "AS-boundary-\(arc4random())-\(arc4random())"
         let type = "multipart/form-data; boundary=\(boundary)"
         let body = self.multiDataFromObject(parameters, multiparts: multiparts, boundary: boundary)
@@ -118,18 +118,16 @@ open class OAuthSwiftClient: NSObject {
         var finalHeaders = [kHTTPHeaderContentType: type]
         finalHeaders += headers ?? [:]
 
-        return makeRequest(urlString, method: method, parameters: parameters, headers: finalHeaders, body: body)
+        return makeRequest(url, method: method, parameters: parameters, headers: finalHeaders, body: body)
     }
 
-    func multiPartRequest(url: String, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters, image: Data, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
+    func multiPartRequest(url: URLConvertible, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters, image: Data, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
         let multiparts = [ OAuthSwiftMultipartData(name: "media", data: image, fileName: "file", mimeType: "image/jpeg") ]
-
-        if let request = makeMultiPartRequest(url, method: method, parameters: parameters, multiparts: multiparts) {
-            request.start(success: success, failure: failure)
-            return request
+        guard let request = makeMultiPartRequest(url, method: method, parameters: parameters, multiparts: multiparts) else {
+            return nil
         }
-
-        return nil
+        request.start(completionHandler: completion)
+        return request
     }
 
     open func multiPartBody(from inputParameters: OAuthSwift.Parameters, boundary: String) -> Data {
@@ -150,15 +148,15 @@ open class OAuthSwiftClient: NSObject {
     }
 
     @discardableResult
-    open func postMultiPartRequest(_ url: String, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, multiparts: [OAuthSwiftMultipartData] = [], checkTokenExpiration: Bool = true, success: OAuthSwiftHTTPRequest.SuccessHandler?, failure: OAuthSwiftHTTPRequest.FailureHandler?) -> OAuthSwiftRequestHandle? {
+    open func postMultiPartRequest(_ url: URLConvertible, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, multiparts: [OAuthSwiftMultipartData] = [], checkTokenExpiration: Bool = true, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) -> OAuthSwiftRequestHandle? {
 
         if checkTokenExpiration && self.credential.isTokenExpired() {
-            failure?(OAuthSwiftError.tokenExpired(error: nil))
+            completion?(.failure(.tokenExpired(error: nil)))
             return nil
         }
 
         if let request = makeMultiPartRequest(url, method: method, parameters: parameters, multiparts: multiparts, headers: headers) {
-            request.start(success: success, failure: failure)
+            request.start(completionHandler: completion)
             return request
         }
         return nil
@@ -191,4 +189,137 @@ open class OAuthSwiftClient: NSObject {
         return data
     }
 
+    // MARK: Refresh Token
+    @discardableResult
+    open func renewAccessToken(accessTokenUrl: URLConvertible?, withRefreshToken refreshToken: String, parameters: OAuthSwift.Parameters? = nil, headers: OAuthSwift.Headers? = nil, contentType: String? = nil, accessTokenBasicAuthentification: Bool = false, completionHandler completion: @escaping OAuthSwift.TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
+        // The current access token isn't needed anymore.
+        self.credential.oauthToken = ""
+
+        var parameters = parameters ?? OAuthSwift.Parameters()
+        parameters["client_id"] = self.credential.consumerKey
+        parameters["refresh_token"] = refreshToken
+        parameters["grant_type"] = "refresh_token"
+
+        // Omit the consumer secret if it's empty; this makes token renewal consistent with PKCE authorization.
+        if !self.credential.consumerSecret.isEmpty {
+            parameters["client_secret"] = self.credential.consumerSecret
+        }
+
+        OAuthSwift.log?.trace("Renew access token, parameters: \(parameters)")
+        return requestOAuthAccessToken(accessTokenUrl: accessTokenUrl, withParameters: parameters, headers: headers, contentType: contentType, accessTokenBasicAuthentification: accessTokenBasicAuthentification, completionHandler: completion)
+    }
+
+    func requestOAuthAccessToken(accessTokenUrl: URLConvertible?, withParameters parameters: OAuthSwift.Parameters, headers: OAuthSwift.Headers? = nil, contentType: String? = nil, accessTokenBasicAuthentification: Bool = false, completionHandler completion: @escaping OAuthSwift.TokenCompletionHandler) -> OAuthSwiftRequestHandle? {
+        OAuthSwift.log?.trace("Request Oauth access token ...")
+        let completionHandler: OAuthSwiftHTTPRequest.CompletionHandler = { [weak self] result in
+            guard let this = self else {
+                OAuthSwift.retainError(completion)
+                return
+            }
+            switch result {
+            case .success(let response):
+                OAuthSwift.log?.trace("Oauth access token response ...")
+
+                let responseJSON: Any? = try? response.jsonObject(options: .mutableContainers)
+
+                let responseParameters: OAuthSwift.Parameters
+
+                if let jsonDico = responseJSON as? [String: Any] {
+                    responseParameters = jsonDico
+                } else {
+                    responseParameters = response.string?.parametersFromQueryString ?? [:]
+                }
+
+                guard let accessToken = responseParameters["access_token"] as? String else {
+                    let message = NSLocalizedString("Could not get Access Token", comment: "Due to an error in the OAuth2 process, we couldn't get a valid token.")
+                    OAuthSwift.log?.error("Could not get access token")
+                    completion(.failure(.serverError(message: message)))
+                    return
+                }
+
+                if let refreshToken = responseParameters["refresh_token"] as? String {
+                    this.credential.oauthRefreshToken = refreshToken.safeStringByRemovingPercentEncoding
+                }
+
+                if let expiresIn = responseParameters["expires_in"] as? String, let offset = Double(expiresIn) {
+                    this.credential.oauthTokenExpiresAt = Date(timeInterval: offset, since: Date())
+                } else if let expiresIn = responseParameters["expires_in"] as? Double {
+                    this.credential.oauthTokenExpiresAt = Date(timeInterval: expiresIn, since: Date())
+                }
+
+                this.credential.oauthToken = accessToken.safeStringByRemovingPercentEncoding
+                completion(.success((this.credential, response, responseParameters)))
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+
+        guard let accessTokenUrl = accessTokenUrl else {
+            let message = NSLocalizedString("access token url not defined", comment: "access token url not defined with code type auth")
+            OAuthSwift.log?.error("Access token url not defined")
+            completion(.failure(.configurationError(message: message)))
+            return nil
+        }
+
+        if contentType == "multipart/form-data" {
+            // Request new access token by disabling check on current token expiration. This is safe because the implementation wants the user to retrieve a new token.
+            return self.postMultiPartRequest(accessTokenUrl, method: .POST, parameters: parameters, headers: headers, checkTokenExpiration: false, completionHandler: completionHandler)
+        } else {
+            // special headers
+            var finalHeaders: OAuthSwift.Headers? = headers
+            if accessTokenBasicAuthentification {
+                let authentification = "\(self.credential.consumerKey):\(self.credential.consumerSecret)".data(using: String.Encoding.utf8)
+                if let base64Encoded = authentification?.base64EncodedString(options: Data.Base64EncodingOptions(rawValue: 0)) {
+                    finalHeaders += ["Authorization": "Basic \(base64Encoded)"] as OAuthSwift.Headers
+                }
+            }
+            // Request new access token by disabling check on current token expiration. This is safe because the implementation wants the user to retrieve a new token.
+            return self.request(accessTokenUrl, method: .POST, parameters: parameters, headers: finalHeaders, checkTokenExpiration: false, completionHandler: completionHandler)
+        }
+    }
+
+    open func requestWithAutomaticAccessTokenRenewal(url: URL, method: OAuthSwiftHTTPRequest.Method, parameters: OAuthSwift.Parameters = [:], headers: OAuthSwift.Headers? = nil, contentType: String? = nil, accessTokenBasicAuthentification: Bool = false, accessTokenUrl: URLConvertible, onTokenRenewal: OAuthSwift.TokenRenewedHandler?, completionHandler completion: OAuthSwiftHTTPRequest.CompletionHandler?) {
+        self.request(url, method: method, parameters: parameters, headers: headers) { [weak self] result in
+            guard let this = self else {
+                OAuthSwift.retainError(completion)
+                return
+            }
+
+            switch result {
+            case .success(let response):
+                if let completion = completion {
+                    completion(.success(response))
+                }
+
+            case .failure(let error):
+                switch error {
+                case OAuthSwiftError.tokenExpired:
+                    if let onTokenRenewal = onTokenRenewal {
+                        let renewCompletionHandler: OAuthSwift.TokenCompletionHandler = { result in
+                            switch result {
+                            case .success(let (credential, _, _)):
+                                onTokenRenewal(.success(credential))
+                                this.requestWithAutomaticAccessTokenRenewal(url: url, method: method, parameters: parameters, headers: headers, contentType: contentType, accessTokenBasicAuthentification: accessTokenBasicAuthentification, accessTokenUrl: accessTokenUrl, onTokenRenewal: nil, completionHandler: completion)
+                            case .failure(let error):
+                                if let completion = completion {
+                                    completion(.failure(.tokenExpired(error: error)))
+                                }
+                            }
+                        }
+
+                        _ = this.renewAccessToken(accessTokenUrl: accessTokenUrl, withRefreshToken: this.credential.oauthRefreshToken, headers: headers, contentType: contentType, accessTokenBasicAuthentification: accessTokenBasicAuthentification, completionHandler: renewCompletionHandler)
+                    } else {
+                        if let completion = completion {
+                            completion(.failure(.tokenExpired(error: nil)))
+                        }
+                    }
+
+                default:
+                    if let completion = completion {
+                        completion(.failure(.tokenExpired(error: nil)))
+                    }
+                }
+            }
+        }
+    }
 }
