@@ -13,8 +13,6 @@ public enum OAuthSwiftError: Error {
 
     /// Configuration problem with oauth provider.
     case configurationError(message: String)
-    /// The provided token is expired, retrieve new token by using the refresh token
-    case tokenExpired(error: Error?)
     /// State missing from request (you can set allowMissingStateCheck = true to ignore)
     case missingState
     /// Returned state value is wrong
@@ -23,17 +21,25 @@ public enum OAuthSwiftError: Error {
     case serverError(message: String)
     /// Failed to create URL \(urlString) not convertible to URL, please encode
     case encodingError(urlString: String)
-    case authorizationPending
     /// Failed to create request with \(urlString)
     case requestCreation(message: String)
     /// Authentification failed. No token
     case missingToken
     /// Please retain OAuthSwift object or handle
     case retain
-    /// Request error
-    case requestError(error: Error, request: URLRequest)
     /// Request cancelled
     case cancelled
+
+    /// Generic request error
+    case requestError(error: Error, request: URLRequest)
+    /// The provided token is expired, retrieve new token by using the refresh token
+    case tokenExpired(error: Error?)
+    /// If the user has not either allowed or denied the request yet, the authorization server will return the authorization_pending error.
+    case authorizationPending(error: Error, request: URLRequest)
+    /// If the device is polling too frequently, the authorization server will return the slow_down error.
+    case slowDown(error: Error, request: URLRequest)
+    /// If the user denies the request.
+    case accessDenied(error: Error, request: URLRequest)
 
     public static let Domain = "OAuthSwiftError"
     public static let ResponseDataKey = "OAuthSwiftError.response.data"
@@ -52,6 +58,8 @@ public enum OAuthSwiftError: Error {
         case retain = -10
         case requestError = -11
         case cancelled = -12
+        case slowDown = -13
+        case accessDenied = -14
     }
 
     fileprivate var code: Code {
@@ -62,12 +70,14 @@ public enum OAuthSwiftError: Error {
         case .stateNotEqual: return Code.stateNotEqual
         case .serverError: return Code.serverError
         case .encodingError: return Code.encodingError
-        case .authorizationPending: return Code.authorizationPending
+        case .cancelled : return Code.cancelled
         case .requestCreation: return Code.requestCreation
         case .missingToken: return Code.missingToken
         case .retain: return Code.retain
         case .requestError: return Code.requestError
-        case .cancelled : return Code.cancelled
+        case .authorizationPending: return Code.authorizationPending
+        case .slowDown: return Code.slowDown
+        case .accessDenied: return Code.accessDenied
         }
     }
 
@@ -75,6 +85,9 @@ public enum OAuthSwiftError: Error {
         switch self {
         case .tokenExpired(let e): return e
         case .requestError(let e, _): return e
+        case .authorizationPending(let e, _): return e
+        case .slowDown(let e, _): return e
+        case .accessDenied(let e, _): return e
         default: return nil
         }
     }
@@ -100,24 +113,29 @@ extension OAuthSwiftError: CustomStringConvertible {
         case .stateNotEqual(let s, let e): return "stateNotEqual[\(s)<>\(e)]"
         case .serverError(let m): return "serverError[\(m)]"
         case .encodingError(let urlString): return "encodingError[\(urlString)]"
-        case .authorizationPending: return "authorizationPending"
         case .requestCreation(let m): return "requestCreation[\(m)]"
         case .missingToken: return "missingToken"
         case .retain: return "retain"
         case .requestError(let e, _): return "requestError[\(e)]"
+        case .slowDown : return "slowDown"
+        case .accessDenied : return "accessDenied"
+        case .authorizationPending: return "authorizationPending"
         case .cancelled : return "cancelled"
         }
     }
 }
 
 extension OAuthSwift {
-
-    static func retainError(_ failureHandler: FailureHandler?) {
+    static func retainError(_ completionHandler: OAuthSwiftHTTPRequest.CompletionHandler?) {
         #if !OAUTH_NO_RETAIN_ERROR
-            failureHandler?(OAuthSwiftError.retain)
+            completionHandler?(.failure(OAuthSwiftError.retain))
         #endif
     }
-
+    static func retainError(_ completionHandler: TokenCompletionHandler?) {
+        #if !OAUTH_NO_RETAIN_ERROR
+        completionHandler?(.failure(OAuthSwiftError.retain))
+        #endif
+    }
 }
 
 // MARK: NSError
@@ -136,6 +154,9 @@ extension OAuthSwiftError: CustomNSError {
 
         case .tokenExpired(let e): return ["error": e as Any]
         case .requestError(let e, let request): return ["error": e, "request": request]
+        case .authorizationPending(let e, let request): return ["error": e, "request": request]
+        case .slowDown(let e, let request): return ["error": e, "request": request]
+        case .accessDenied(let e, let request): return ["error": e, "request": request]
 
         case .encodingError(let urlString): return ["url": urlString]
 
