@@ -7,18 +7,25 @@
 //
 
 import Foundation
+#if os(iOS)
+#if !OAUTH_APP_EXTENSIONS
+import UIKit
+#endif
+#endif
 
 let kHTTPHeaderContentType = "Content-Type"
 
 open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
 
-    public typealias SuccessHandler = (_ response: OAuthSwiftResponse) -> Void
-    public typealias FailureHandler = (_ error: OAuthSwiftError) -> Void
+    // Using NSLock for Linux compatible locking 
+    let requestLock = NSLock()
+
+    public typealias CompletionHandler = (_ result: Result<OAuthSwiftResponse, OAuthSwiftError>) -> Void
 
     /// HTTP request method
     /// https://en.wikipedia.org/wiki/Hypertext_Transfer_Protocol#Request_methods
     public enum Method: String {
-        case GET, POST, PUT, DELETE, PATCH, HEAD //, OPTIONS, TRACE, CONNECT
+        case GET, POST, PUT, DELETE, PATCH, HEAD // , OPTIONS, TRACE, CONNECT
 
         var isBody: Bool {
             return self == .POST || self == .PUT || self == .PATCH
@@ -57,25 +64,23 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
     }
 
     /// START request
-    func start(success: SuccessHandler?, failure: FailureHandler?) {
+    func start(completionHandler completion: CompletionHandler?) {
         guard request == nil else { return } // Don't start the same request twice!
-
-        let successHandler = success
-        let failureHandler = failure
 
         do {
             self.request = try self.makeRequest()
         } catch let error as NSError {
-            failureHandler?(OAuthSwiftError.requestCreation(message: error.localizedDescription))
+            completion?(.failure(.requestCreation(message: error.localizedDescription)))
             self.request = nil
             return
         }
 
         OAuthSwiftHTTPRequest.executionContext {
             // perform lock here to prevent cancel calls on another thread while creating the request
-            objc_sync_enter(self)
-            defer { objc_sync_exit(self) }
+            self.requestLock.lock()
+            defer { self.requestLock.unlock() }
             if self.cancelRequested {
+                completion?(.failure(.cancelled))
                 return
             }
 
@@ -84,8 +89,7 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
 
             if self.config.sessionFactory.useDataTaskClosure {
                 let completionHandler: (Data?, URLResponse?, Error?) -> Void = { data, resp, error in
-                    OAuthSwiftHTTPRequest.completionHandler(successHandler: success,
-                                                            failureHandler: failure,
+                    OAuthSwiftHTTPRequest.completionHandler(completionHandler: completion,
                                                             request: usedRequest,
                                                             data: data,
                                                             resp: resp,
@@ -101,17 +105,21 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
 
             #if os(iOS)
                 #if !OAUTH_APP_EXTENSIONS
+                #if !targetEnvironment(macCatalyst)
                     UIApplication.shared.isNetworkActivityIndicatorVisible = self.config.sessionFactory.isNetworkActivityIndicatorVisible
+                    #endif
                 #endif
             #endif
         }
     }
 
     /// Function called when receiving data from server.
-    public static func completionHandler(successHandler: SuccessHandler?, failureHandler: FailureHandler?, request: URLRequest, data: Data?, resp: URLResponse?, error: Error?) {
+    public static func completionHandler(completionHandler completion: CompletionHandler?, request: URLRequest, data: Data?, resp: URLResponse?, error: Error?) {
         #if os(iOS)
         #if !OAUTH_APP_EXTENSIONS
+        #if !targetEnvironment(macCatalyst)
         UIApplication.shared.isNetworkActivityIndicatorVisible = false
+        #endif
         #endif
         #endif
 
@@ -125,7 +133,7 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
                 oauthError = .tokenExpired(error: error)
             }
 
-            failureHandler?(oauthError)
+            completion?(.failure(oauthError))
             return
         }
 
@@ -143,24 +151,25 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
                 userInfo["Response-Headers"] = response.allHeaderFields
             }
             let error = NSError(domain: OAuthSwiftError.Domain, code: badRequestCode, userInfo: userInfo)
-            failureHandler?(.requestError(error:error, request: request))
+            completion?(.failure(.requestError(error:error, request: request)))
             return
         }
 
         // MARK: failure code > 400
         guard response.statusCode < 400 else {
-            var localizedDescription = String()
+            var localizedDescription = ""
             let responseString = String(data: responseData, encoding: OAuthSwiftDataEncoding)
 
             // Try to get error information from data as json
             let responseJSON = try? JSONSerialization.jsonObject(with: responseData, options: .mutableContainers)
+            var errorCode: String?
             if let responseJSON = responseJSON as? OAuthSwift.Parameters {
-                if let code = responseJSON["error"] as? String, let description = responseJSON["error_description"] as? String {
-
-                    localizedDescription = NSLocalizedString("\(code) \(description)", comment: "")
-                    if code == "authorization_pending" {
-                        failureHandler?(.authorizationPending)
-                        return
+                if let code = responseJSON["error"] as? String {
+                    errorCode = code
+                    if  let description = responseJSON["error_description"] as? String {
+                        localizedDescription = NSLocalizedString("\(code) \(description)", comment: "")
+                    } else {
+                        localizedDescription = NSLocalizedString("\(code)", comment: "")
                     }
                 }
             } else {
@@ -180,23 +189,29 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
                 userInfo[NSURLErrorFailingURLErrorKey] = urlString
             }
 
-            let error = NSError(domain: NSURLErrorDomain, code: response.statusCode, userInfo: userInfo)
+            let error = NSError(domain: OAuthSwiftError.Domain, code: response.statusCode, userInfo: userInfo)
             if error.isExpiredToken {
-                failureHandler?(.tokenExpired(error: error))
+                completion?(.failure(.tokenExpired(error: error)))
+            } else if errorCode == "authorization_pending" {
+                completion?(.failure(.authorizationPending(error: error, request: request)))
+            } else if errorCode == "slow_down" {
+                completion?(.failure(.slowDown(error: error, request: request)))
+            } else if errorCode == "access_denied" {
+                completion?(.failure(.accessDenied(error: error, request: request)))
             } else {
-                failureHandler?(.requestError(error: error, request: request))
+                completion?(.failure(.requestError(error: error, request: request)))
             }
             return
         }
 
         // MARK: success
-        successHandler?(OAuthSwiftResponse(data: responseData, response: response, request: request))
+        completion?(.success(OAuthSwiftResponse(data: responseData, response: response, request: request)))
     }
 
     open func cancel() {
         // perform lock here to prevent cancel calls on another thread while creating the request
-        objc_sync_enter(self)
-        defer { objc_sync_exit(self) }
+        requestLock.lock()
+        defer { requestLock.unlock() }
         // either cancel the request if it's already running or set the flag to prohibit creation of the request
         if let task = task {
             task.cancel()
@@ -211,6 +226,7 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
 
     open class func makeRequest(config: Config) throws -> URLRequest {
         var request = config.urlRequest
+        OAuthSwift.log?.trace("URLRequest is created: \(request)")
         return try setupRequestForOAuth(request: &request,
                                         parameters: config.parameters,
                                         dataEncoding: config.dataEncoding,
@@ -232,6 +248,7 @@ open class OAuthSwiftHTTPRequest: NSObject, OAuthSwiftRequestHandle {
         for (key, value) in headers {
             request.setValue(value, forHTTPHeaderField: key)
         }
+        OAuthSwift.log?.trace("URLRequest is created: \(request)")
 
         return try setupRequestForOAuth(
             request: &request,
@@ -396,10 +413,10 @@ extension OAuthSwiftHTTPRequest {
             var requestHeaders = OAuthSwift.Headers()
             switch paramsLocation {
             case .authorizationHeader:
-                //Add oauth parameters in the Authorization header
+                // Add oauth parameters in the Authorization header
                 requestHeaders += credential.makeHeaders(signatureUrl, method: method, parameters: signatureParameters, body: body)
             case .requestURIQuery:
-                //Add oauth parameters as request parameters
+                // Add oauth parameters as request parameters
                 self.parameters += credential.authorizationParametersWithSignature(method: method, url: signatureUrl, parameters: signatureParameters, body: body)
             }
 
